requirements.txt
dbt-snowflake==1.7.1
snowflake-connector-python==3.6.0
toml==0.10.2
